feat: 이미지 리스트 조회 및 관리 기능 구현

+++ file-service/src/main/java/com/momnect/fileservice/command/repository/ImageFileRepository.java
@@ -3,5 +3,5 @@ package com.momnect.fileservice.command.repository;
 import com.momnect.fileservice.command.entity.ImageFile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImageFileRepository extends JpaRepository<ImageFile, Integer> {
+public interface ImageFileRepository extends JpaRepository<ImageFile, Long> {
 }
